refactor: extract ObjectDecoders from MapDecoders and JooqRecordDecoders
Primitive decoders (string, int_, bool, long_, decimal, temporal types,
nullable, list, map, enumOf, literal) were duplicated in both MapDecoders
and JooqRecordDecoders. Extract them into a new ObjectDecoders class in
the net.unit8.raoh top-level package so they can be reused in custom
decoder classes (e.g. FormDecoders) without depending on any boundary
module.

Also bumps version to 0.4.0-SNAPSHOT for the next development cycle.
