+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>net.unit8.raoh</groupId>
     <artifactId>raoh-parent</artifactId>
-    <version>0.3.1</version>
+    <version>0.4.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <name>Raoh Parent</name>
