+++ raoh/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>net.unit8.raoh</groupId>
         <artifactId>raoh-parent</artifactId>
-        <version>0.3.1</version>
+        <version>0.4.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
