+++ raoh/src/main/java/net/unit8/raoh/builtin/package-info.java
@@ -2,6 +2,7 @@
  * Built-in typed decoders with fluent constraint APIs.
  *
  * <p>These decoders are typically created via factory methods in
- * {@code net.unit8.raoh.json.JsonDecoders} or {@link net.unit8.raoh.map.MapDecoders}.
+ * {@link net.unit8.raoh.ObjectDecoders}, {@code net.unit8.raoh.json.JsonDecoders},
+ * or {@link net.unit8.raoh.map.MapDecoders}.
  */
 package net.unit8.raoh.builtin;
