+++ raoh/src/main/java/net/unit8/raoh/map/package-info.java
@@ -1,5 +1,10 @@
 /**
- * Decoders for {@code Map<String, Object>} input.
+ * Decoders for {@code Map<String, Object>} structure concerns.
+ *
+ * <p>Provides field extraction ({@code field()}, {@code optionalField()}, etc.),
+ * applicative combination ({@code combine()}), and structural validation ({@code strict()}).
+ * For primitive value decoders ({@code string()}, {@code int_()}, etc.),
+ * see {@link net.unit8.raoh.ObjectDecoders}.
  *
  * @see net.unit8.raoh.map.MapDecoders
  */
