+++ raoh/src/main/java/net/unit8/raoh/package-info.java
@@ -11,8 +11,9 @@
  *
  * <p>For input-specific factories, see:
  * <ul>
+ *   <li>{@link net.unit8.raoh.ObjectDecoders} — for raw {@code Object} values (field values from maps, jOOQ records, etc.)</li>
  *   <li>{@code net.unit8.raoh.json.JsonDecoders} — for Jackson {@code JsonNode}</li>
- *   <li>{@link net.unit8.raoh.map.MapDecoders} — for {@code Map<String, Object>}</li>
+ *   <li>{@link net.unit8.raoh.map.MapDecoders} — for {@code Map<String, Object>} structure (field extraction, combine)</li>
  * </ul>
  */
 package net.unit8.raoh;
