+++ raoh/src/test/java/net/unit8/raoh/CombineListTest.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Map;
 
 import static net.unit8.raoh.map.MapDecoders.*;
+import static net.unit8.raoh.ObjectDecoders.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
